\name{bivrec}
\alias{bivrec}
\alias{bivrec.formula}
\alias{bivrec.data.frame}
\alias{print.bivrec}
\alias{bivrec.object}
\title{ Survival analysis for bivariate recurrent event processes
     with BLUP frailties }
\description{
 Tools for fitting proportional hazards models to clustered
 bivariate recurrent events data. Nested frailties are modeled by their
 best linear unbiased predictors under an auxiliary Poisson model.
}
\usage{\method{bivrec}{formula}(formula, data = parent.frame(), ...)}
\arguments{
    \item{formula}{See details}
    \item{data}{See details}
    \item{\dots}{See details}
}
\details{
    Supported parameters are:
    \preformatted{bivrec(formula, data = parent.frame(), K1 = 10, K2 = 10,
        excludevars1 = NULL, excludevars2 = NULL, verbose = 1, 
        alternating = FALSE, dispest = "pearson", correction = "none", 
        computesd = TRUE, fullS = TRUE, fixzero = NULL, smooth = FALSE, 
        maxiter = 200, convergence = 1e-3, initial = NULL)}
   \describe{
   \item{formula}{a formula object, similar to \code{\link{coxph}}. The response to the
       left of the \code{~} should be a survival object generated by \code{\link{Surv2}}. The
       right side must contain an \code{id(x)} term, where \code{x} is a variable that takes a
       unique value for each subject within a cluster. It may also contain a \code{cluster(y)} 
       term if \code{y} is the variable that indicates cluster membership, and a \code{strata(z)}
       term.
   }
   \item{data}{a \code{data.frame} with columns corresponding to the terms 
       in the formula.}
  \item{K1}{either an integer, a vector of integers, or a value between 0 and 1,
    to determine the level of discretization. If it is an integer, \code{K1} gives the
    number of breakpoints in the baseline hazard for the first process. 
    If it is a vector of integers, \code{K1} should have length equal to the 
    number of strata, and each value gives the number of breakpoints in the 
    baseline hazard for each stratum. If it is a number between 0 and 1, it gives
    the ratio of the number of breakpoints relative to the maximum possible. 
    Defaults to 10.}
  \item{K2}{ analogous to \code{K1}, for the second process. }
  \item{excludevars1}{ a vector of strings giving the names of variables or 
    interactions that should be excluded from the model for the first process. }
  \item{excludevars2}{ analogous to \code{excludevars1} for the second process. }
  \item{verbose}{ an integer from 0 to 3 that determines the quantity of output
        printed to the screen. Setting \code{verbose=0} is completely silent.}
  \item{alternating}{ logical, describing the at-risk function for the model. If
    \code{FALSE}, patients are assumed to be constantly at risk for both processes,
    if \code{TRUE}, they are only at-risk for one process at a time. Defaults to 
    \code{FALSE}. }
  \item{dispest}{ a string, determining the method used to estimate the dispersion
    parameters. Possible values are \code{"pearson", "marginal", "ohlsson"}, defaults
    to \code{"pearson"}. }
  \item{correction}{ a string describing the degree-of-freedom correction proposed
    by Ma. It only applies when \code{dispest="pearson"}. 
    Possible values are \code{"single", "double", "none"}, defaults to \code{"none"} }
  \item{computesd}{ a boolean determining whether standard errors should be computed. 
    Defaults to \code{TRUE}. }
  \item{fullS}{ logical determining whether to use the full sensitivity matrix
    in computing standard errors, or only the covariate portion. Using \code{fullS=TRUE}
    leads to more accurate results but increases computer time. Defaults to \code{TRUE}. }
  \item{fixzero}{ a vector of strings, listing any dispersion parameters that should
    be set to 0. Can contain any subset of \code{"clust1", "clust2", "subj1", "subj2", "cov"},
    all else is ignored. Defaults to \code{NULL}. }
  \item{smooth}{ logical, determines whether the baseline hazard should be smoothed
    at each iteration. Defaults to \code{FALSE}. }
  \item{maxiter}{ an integer giving the maximum number of iterations permitted. }
  \item{convergence}{ double, determines the value of the convergence criterion 
    required at termination. }
  \item{initial}{ a list of initial values in the format used by the code internally. }
  }
}
\value{
     An object of class \code{bivrec} with the following components:
     \item{call}{the original call to the model-fitting function}
     \item{regression}{a list containing results from the regression fit in the last
        iteration. It has components
        \describe{
            \item{\code{coefficients1}}{a vector of regression coefficients for the
                first process}
            \item{\code{coefficients2}}{a vector of regression coefficients for the
                second process}            
            \item{\code{loglik1}}{conditional loglikelihood for the first process.}
            \item{\code{loglik2}}{conditional loglikelihood for the second process.}       
        }
     }
     \item{frailty}{a list containing results from the frailty estimation in the last
        iteration. It has components
        \describe{
            \item{\code{clust1}}{cluster frailties for process 1}
            \item{\code{clust2}}{cluster frailties for process 2}            
            \item{\code{subj1}}{subject frailties for process 1}
            \item{\code{subj2}}{subject frailties for process 2}       
        }
     }
     \item{dispersion}{a list containing results from the dispersion parameter
         estimation in the last iteration. It has components
        \describe{
            \item{\code{clust1}}{cluster frailty variance for process 1}
            \item{\code{clust2}}{cluster frailty variance for process 2}            
            \item{\code{subj1}}{subject frailty variance for process 1}
            \item{\code{subj2}}{subject frailty variance for process 2}       
            \item{\code{cov}}{subject frailty covariance}       
        }
     }
     \item{hazard}{a list describing the baseline hazard. It has components
        \describe{
            \item{\code{breaks1}}{matrix of breakpoints in the hazard for 
                each stratum for process 1}
            \item{\code{breaks2}}{matrix of breakpoints in the hazard for 
                each stratum for process 2}
            \item{\code{hazard1}}{matrix of hazards in each interval for
                each stratum for process 1}
            \item{\code{hazard2}}{matrix of hazards in each interval for
                each stratum for process 1}
        }
     }
     \item{summaries}{a list of summary matrices. It has components
        \describe{
            \item{\code{regression}}{summary for the regression coefficients}
            \item{\code{dispersion}}{summary for the dispersion parameters}
        }
     }

}
\references{E. Sharef and R. Strawderman. \dQuote{A nested frailty model for
clustered bivariate recurrent events}, \emph{in preparation}.}
\author{Emmanuel Sharef \email{ess28@cornell.edu} }
\seealso{\code{\link{summary.bivrec}}, \code{\link{plot.bivrec}}}
\examples{
    data(m10Ji5)
    
    fit <- bivrec( Surv2(start, stop, delta, Delta) ~ 
                Z1 + cluster(i) + id(j), data = m10Ji5 )
                
    summary(fit)
    plot(fit)
}
\keyword{ survival }
