\name{id}
\alias{id}
\title{ Identify subjects }
\description{
 A function used in the context of specifying recurrent event models to identify
 individual subjects.
}
\usage{
id(x)
}
\arguments{
  \item{x}{a character, factor or numeric variable that should be unique for
    each subject (or each subject within a cluster).}
}
\value{
    \code{x}
}
\seealso{ \code{\link{Surv2}}, \code{\link{bivrec}}}
\keyword{ survival }
